refactor: update import paths for schemas and create new schema files for comments, players, seasons, and songs
