+++ apps/assassin/src/features/song/repository.ts
@@ -1,5 +1,5 @@
 import { supabase } from "@libs/supabase/client";
-import { CreateSongInput, UpdateSongInput } from "./schemas/song";
+import { CreateSongInput, UpdateSongInput } from "./schema";
 
 async function getAllSongs() {
   const { data, error } = await supabase
